# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduAPP API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/event/{type}/{id}/{time}": {
            "get": {
                "tags": [
                    "Attandances"
                ],
                "summary": "student and staff attandance",
                "description": "student and staff attandance",
                "operationId": "attandanceEvent",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "type should be 'staff' or 'student' ",
                        "required": true,
                        "example": "student"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "attandance person ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "time",
                        "in": "path",
                        "description": "attandance time",
                        "required": true,
                        "example": "09:00"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Student or Staff not found "
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost:8000/oauth/authorize",
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "refreshUrl": "http://localhost:8000/token/refresh",
                        "scopes": []
                    }
                }
            },
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}